requirements.txt
opencv-python>=4.5.0
opencv-contrib-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
streamlit>=1.28.0
pillow>=9.0.0
